+++ routing/http/contentrouter/contentrouter.go
@@ -119,7 +119,7 @@ func readProviderResponses(ctx context.Context, iter iter.ResultIter[types.Recor
 	for iter.Next() {
 		res := iter.Val()
 		if res.Err != nil {
-			logger.Warnw("error iterating provider responses: %s", res.Err)
+			logger.Warnf("error iterating provider responses: %s", res.Err)
 			continue
 		}
 		v := res.Val
@@ -200,21 +200,32 @@ func (c *contentRouter) FindPeer(ctx context.Context, pid peer.ID) (peer.AddrInf
 	for iter.Next() {
 		res := iter.Val()
 		if res.Err != nil {
-			logger.Warnw("error iterating provider responses: %s", res.Err)
+			logger.Warnf("error iterating peer responses: %s", res.Err)
 			continue
 		}
+
+		if *res.Val.ID != pid {
+			logger.Warnf("searched for peerID %s, got response for %s:", pid, *res.Val.ID)
+			continue
+		}
+
 		var addrs []multiaddr.Multiaddr
 		for _, a := range res.Val.Addrs {
 			addrs = append(addrs, a.Multiaddr)
 		}
 
+		// If there are no addresses there's nothing of value to return
+		if len(addrs) == 0 {
+			continue
+		}
+
 		return peer.AddrInfo{
-			ID:    *res.Val.ID,
+			ID:    pid,
 			Addrs: addrs,
 		}, nil
 	}
 
-	return peer.AddrInfo{}, err
+	return peer.AddrInfo{}, routing.ErrNotFound
 }
 
 func (c *contentRouter) PutValue(ctx context.Context, key string, data []byte, opts ...routing.Option) error {